move ivxplayer to tempo.py
